задание 7/requirements.txt
yandex-cloud-ml-sdk>=0.2.0
python-telegram-bot>=20.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
